fix(zh): preserve land acknowledgement scope
Translate each infrastructure relationship as operational rather than constructional, with a focused fidelity guard. Move exhaustive localized Blog and Glossary navigation matrices into a dedicated integration spec so smoke coverage stays lean.

中文：将三处基础设施与传统领地的关系准确表达为“运行于”，避免误写成建设关系，并加入聚焦语义范围的回归测试；同时把 Blog 与 Glossary 的完整双语导航矩阵移入独立集成测试，保持 smoke 测试轻量。
